requirements.txt
# Medical Form Template Extraction System
# Phase 1: Scan blank PDF forms -> Generate Pydantic templates/schemas

# Core dependencies
instructor>=1.7.0
openai>=1.0.0
anthropic>=0.40.0
pydantic>=2.5.0

# PDF processing
pdf2image>=1.17.0
Pillow>=10.0.0

# CLI
click>=8.1.0

# Environment
python-dotenv>=1.0.0

# Utilities
rich>=13.0.0  # Pretty console output

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6  # File upload handling

# Database & Storage
supabase>=2.0.0

# Auth
PyJWT>=2.8.0
